requirements.txt
aiohttp==3.9.3
openai>=1.12.0
python-dotenv>=0.19.0
requests==2.31.0
python-telegram-bot==20.3
